Table refactor P2
